+++ src/actions/actions.ts
@@ -1,0 +1,16 @@
+"use server"
+
+export async function generateImage( prevState: State,formData:FormData){
+    const keyword = formData.get("keyword");
+    try {
+        await fetch(`${process.env.BASE_URL}/api/generate-image`,{
+            method:"POST",
+            headers:{
+                "Content-Type": "application/json",
+            },
+            body: JSON.stringify({keyword}),
+        })
+    } catch (error) {
+        console.log(error)
+    }
+}
